Unit project 5 complete

+++ Application/src/main/java/com/kenzie/marketing/application/service/CustomerService.java
@@ -14,9 +14,12 @@ import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
+import java.sql.Ref;
 import java.time.LocalDateTime;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
+import java.util.Random;
 import java.util.stream.Collectors;
 import java.util.stream.StreamSupport;
 
@@ -45,7 +48,9 @@ public class CustomerService {
 
         // Task 1 - Add your code here
 
-        return null;
+        return records.stream()
+                .map(r -> toCustomerResponse(r))
+                .collect(Collectors.toList());
     }
 
     /**
@@ -56,9 +61,12 @@ public class CustomerService {
     public CustomerResponse getCustomer(String customerId) {
         Optional<CustomerRecord> record = customerRepository.findById(customerId);
 
-        // Task 1 - Add your code here
+        if (record.isEmpty()) {
+            return null;
+        }
 
-        return null;
+        // Task 1 - Add your code here
+        return toCustomerResponse(record.get());
     }
 
     /**
@@ -71,9 +79,35 @@ public class CustomerService {
      */
     public CustomerResponse addNewCustomer(CreateCustomerRequest createCustomerRequest) {
 
-        // Task 1 - Add your code here
+        if (createCustomerRequest.getReferrerId().isPresent() && createCustomerRequest.getReferrerId().get().length() == 0) {
+            createCustomerRequest.setReferrerId(Optional.empty());
+        }
 
-        return null;
+        if (createCustomerRequest.getReferrerId().isPresent()) {
+            if (!customerRepository.existsById(createCustomerRequest.getReferrerId().get())) {
+                throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "ID does not exist");
+            }
+        }
+
+        CustomerRecord customerRecord = new CustomerRecord();
+        customerRecord.setId(randomUUID().toString());
+        customerRecord.setName(createCustomerRequest.getName());
+        customerRecord.setDateCreated(LocalDateTime.now().toString());
+        customerRecord.setReferrerId(createCustomerRequest.getReferrerId().orElse(null));
+
+        ReferralRequest referralRequest = new ReferralRequest();
+        referralRequest.setCustomerId(customerRecord.getId());
+        referralRequest.setReferrerId(customerRecord.getReferrerId());
+
+        customerRepository.save(customerRecord);
+
+        if (createCustomerRequest.getReferrerId().isPresent() && createCustomerRequest.getReferrerId().get().length() == 0) {
+            referralServiceClient.addReferral(new ReferralRequest(customerRecord.getId(), Optional.empty().toString()));
+        }
+
+        referralServiceClient.addReferral(referralRequest);
+
+        return toCustomerResponse(customerRecord);
     }
 
     /**
@@ -91,8 +125,7 @@ public class CustomerService {
         customerRepository.save(customerRecord);
 
         // Task 1 - Add your code here
-
-        return null;
+        return toCustomerResponse(customerRecord);
     }
 
     /**
@@ -100,6 +133,9 @@ public class CustomerService {
      * @param customerId
      */
     public void deleteCustomer(String customerId) {
+        if (customerId == null) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Invalid Customer Id");
+        }
         customerRepository.deleteById(customerId);
     }
 
@@ -127,9 +163,24 @@ public class CustomerService {
      */
     public List<CustomerResponse> getReferrals(String customerId) {
 
-        // Task 1 - Add your code here
+        
+        if (customerId == null) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "ID does not exist");
+        }
 
-        return null;
+        // Task 1 - Add your code here
+        CustomerRecord record = customerRepository.findById(customerId)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.BAD_REQUEST, "Customer Not Found"));
+
+        CustomerRecord referralRecord = new CustomerRecord();
+
+        return  referralServiceClient.getDirectReferrals(customerId).stream()
+                .peek(r -> referralRecord.setReferrerId(r.getReferrerId()))
+                .peek(r -> referralRecord.setId(r.getCustomerId()))
+                .peek(r -> referralRecord.setDateCreated(r.getReferralDate()))
+                .peek(r -> referralRecord.setName(record.getName()))
+                .map(r -> toCustomerResponse(referralRecord))
+                .collect(Collectors.toList());
     }
 
     /**
@@ -140,7 +191,18 @@ public class CustomerService {
 
         // Task 2 - Add your code here
 
-        return null;
+        List<LeaderboardEntry> board = referralServiceClient.getLeaderboard();
+        List<LeaderboardUiEntry> uiEntries = new ArrayList<>();
+
+        for (LeaderboardEntry entry : board) {
+            LeaderboardUiEntry uiEntry = new LeaderboardUiEntry();
+            uiEntry.setCustomerId(entry.getCustomerId());
+            uiEntry.setCustomerName("No name found");
+            uiEntry.setNumReferrals(entry.getNumReferrals());
+            uiEntries.add(uiEntry);
+        }
+
+        return uiEntries;
     }
 
     /* -----------------------------------------------------------------------------------------------------------
@@ -149,4 +211,24 @@ public class CustomerService {
 
     // Add any private methods here
 
+    private CustomerResponse toCustomerResponse(CustomerRecord record) {
+        if (record == null) {
+            return null;
+        }
+
+        CustomerResponse customerResponse = new CustomerResponse();
+
+        if (record.getReferrerId() != null && !record.getReferrerId().isEmpty()) {
+            if (customerRepository.findById(record.getReferrerId()).isPresent()) {
+                customerResponse.setReferrerName(customerRepository.findById(record.getReferrerId()).get().getName());
+            }
+        }
+
+        customerResponse.setId(record.getId());
+        customerResponse.setName(record.getName());
+        customerResponse.setDateJoined(record.getDateCreated());
+        customerResponse.setReferrerId(record.getReferrerId());
+
+        return customerResponse;
+    }
 }
